## Summary
  - enable stricter debug warning handling for C/C++ and CUDA builds
- remove the sign-compare suppression by fixing the underlying warnings

+++ gsplat/cuda/csrc/CameraWrappers.cu
@@ -30,9 +30,6 @@
 
 namespace gsplat {
 /**
- * @brief Reshapes a tensor to match a specified shape template.
- *
-/**
  * @brief Normalizes the shape of an input tensor according to a specified shape template.
  *
  * This function adjusts the dimensions of the provided tensor to match the number of dimensions
@@ -57,7 +54,7 @@ torch::Tensor normalize_shape(torch::Tensor tensor, std::string_view tensor_name
         // Reshape by prepending 1's if too few dimensions
         if(tensor.dim() < expected_ndims)
         {
-            tensor = tensor.expand({((Is < expected_ndims - tensor.dim()) ? 1 : -1)...});
+            tensor = tensor.expand({((static_cast<int>(Is) < expected_ndims - static_cast<int>(tensor.dim())) ? 1 : -1)...});
         }
 
         // Flatten leftmost dimensions if too many, or use as-is
